+++ src/lungo_cli/plugins/stirlingpdf/web/routes/[...url]/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+    import { PatchedIFrame } from "$lib/components"
+</script>
+
+<PatchedIFrame />